# next/service/service-api/src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WePush Next Service API
  version: 0.1.0
  description: Local-first control plane API shared by Web UI, Desktop UI and remote SDKs.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workspaces:
    get:
      operationId: listWorkspaces
      tags: [Security]
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Workspaces, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Workspace'}}}}}
    post:
      operationId: createWorkspace
      tags: [Security]
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateWorkspace'}}}}
      responses:
        '201': {description: Workspace created, content: {application/json: {schema: {$ref: '#/components/schemas/Workspace'}}}}
  /api/v1/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [Security]
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
      responses:
        '200': {description: Workspace, content: {application/json: {schema: {$ref: '#/components/schemas/Workspace'}}}}
...
  /api/v1/workspaces/{workspaceId}/schedules:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    get:
      operationId: listSchedules
      tags: [Schedules]
      responses:
        '200':
          description: Workspace schedules
          content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Schedule'}}}}
    post:
      operationId: createSchedule
      tags: [Schedules]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateSchedule'}}}
      responses:
        '201':
          description: Schedule created
          content: {application/json: {schema: {$ref: '#/components/schemas/Schedule'}}}
  /api/v1/workspaces/{workspaceId}/schedules/{scheduleId}:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {name: scheduleId, in: path, required: true, schema: {type: string}}
    patch:
      operationId: setScheduleEnabled
      tags: [Schedules]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateSchedule'}}}
      responses:
        '200':
          description: Schedule state updated
          content: {application/json: {schema: {$ref: '#/components/schemas/Schedule'}}}
    delete:
      operationId: deleteSchedule
      tags: [Schedules]
      responses:
        '204': {description: Schedule deleted}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
    enrollmentAuth: {type: http, scheme: Enrollment}
    agentAuth: {type: http, scheme: Agent}
    agentHeaderAuth: {type: apiKey, in: header, name: x-wepush-agent-token}
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {type: string}}
...
  schemas:
...
    CreateSchedule:
      type: object
      required: [name, jobId, cronExpression, timezone, misfirePolicy]
      properties:
        name: {type: string}
        jobId: {type: string}
        cronExpression: {type: string, description: Spring six-field cron expression}
        timezone: {type: string, examples: [Asia/Shanghai]}
        misfirePolicy: {type: string, enum: [FIRE_ONCE, SKIP]}
        enabled: {type: boolean, default: true}
    UpdateSchedule:
      type: object
      required: [enabled]
      properties: {enabled: {type: boolean}}
    Schedule:
      type: object
      required: [id, workspaceId, jobId, name, cronExpression, timezone, misfirePolicy, enabled,
                 nextFireAt, createdAt, updatedAt, version]
      properties:
        id: {type: string}
        workspaceId: {type: string}
        jobId: {type: string}
        name: {type: string}
        cronExpression: {type: string}
        timezone: {type: string}
        misfirePolicy: {type: string, enum: [FIRE_ONCE, SKIP]}
        enabled: {type: boolean}
        nextFireAt: {type: string, format: date-time}
        lastFireAt: {type: [string, 'null'], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        version: {type: integer, format: int64}
...
    CreateWorkspace:
      type: object
      required: [name]
      properties: {name: {type: string}}
    Workspace:
      type: object
      required: [id, name, status, createdAt, version]
      properties:
        id: {type: string}
        name: {type: string}
        status: {type: string, enum: [ACTIVE, SUSPENDED, ARCHIVED]}
        createdAt: {type: string, format: date-time}
        version: {type: integer, format: int64, minimum: 0}
...